requirements.txt
upstox-python-sdk
pandas
numpy
requests
beautifulsoup4
textblob
python-dotenv
pandas-ta
